UfoEditor prepared

+++ UFO Builder/Game.cpp
@@ -105,7 +105,7 @@ void Game::tick(sf::Time elapsedTime)
 			{
 				// left mouse button has been clicked
 				m_lastClick = sf::Mouse::getPosition();
-				m_difference = sf::Vector2i(m_lastClick.x - m_spriteBody.getPosition().x, m_lastClick.y - m_spriteBody.getPosition().y);
+				m_difference = sf::Vector2i(m_lastClick.x - (int)m_spriteBody.getPosition().x, m_lastClick.y - (int)m_spriteBody.getPosition().y);
 
 				if(isSelected(m_spriteBody))
 					m_isPressed = true;
@@ -125,8 +125,8 @@ void Game::tick(sf::Time elapsedTime)
 
 	if(m_isPressed)
 	{
-		m_spriteBody.setPosition(m_mousePosition.x - m_difference.x, m_mousePosition.y - m_difference.y);
-		m_bodyTest->SetTransform(b2Vec2(m_mousePosition.x, m_mousePosition.y), m_bodyTest->GetAngle());
+		m_spriteBody.setPosition((float)(m_mousePosition.x - m_difference.x), (float)(m_mousePosition.y - m_difference.y));
+		m_bodyTest->SetTransform(b2Vec2((float)(m_mousePosition.x), (float)(m_mousePosition.y)), m_bodyTest->GetAngle());
 	}
 	else
 	{
